requirements.txt
altair                       5.5.0
annotated-types              0.7.0
anyio                        4.10.0
attrs                        25.3.0
beautifulsoup4               4.13.5
blinker                      1.9.0
cachetools                   5.5.2
certifi                      2025.8.3
charset-normalizer           3.4.3
click                        8.2.1
clip                         1.0
colorama                     0.4.6
ddgs                         9.5.5
duckduckgo_search            8.1.1
filelock                     3.19.1
fsspec                       2025.7.0
ftfy                         6.3.1
gitdb                        4.0.12
GitPython                    3.1.45
google-ai-generativelanguage 0.6.15
google-api-core              2.25.1
google-api-python-client     2.179.0
google-auth                  2.40.3
google-auth-httplib2         0.2.0
google-genai                 1.32.0
google-generativeai          0.8.5
googleapis-common-protos     1.70.0
grpcio                       1.74.0
grpcio-status                1.71.2
h11                          0.16.0
httpcore                     1.0.9
httplib2                     0.30.0
httpx                        0.28.1
huggingface-hub              0.34.4
idna                         3.10
Jinja2                       3.1.6
jsonschema                   4.25.1
jsonschema-specifications    2025.4.1
lxml                         6.0.1
MarkupSafe                   3.0.2
mpmath                       1.3.0
narwhals                     2.3.0
networkx                     3.5
numpy                        2.2.6
opencv-python                4.12.0.88
packaging                    25.0
pandas                       2.3.2
pillow                       11.3.0
pip                          25.2
primp                        0.15.0
proto-plus                   1.26.1
protobuf                     5.29.5
pyarrow                      21.0.0
pyasn1                       0.6.1
pyasn1_modules               0.4.2
pydantic                     2.11.7
pydantic_core                2.33.2
pydeck                       0.9.1
pyparsing                    3.2.3
python-dateutil              2.9.0.post0
pytz                         2025.2
PyYAML                       6.0.2
referencing                  0.36.2
regex                        2025.8.29
requests                     2.32.5
rpds-py                      0.27.1
rsa                          4.9.1
safetensors                  0.6.2
setuptools                   65.5.0
six                          1.17.0
smmap                        5.0.2
sniffio                      1.3.1
soupsieve                    2.8
streamlit                    1.49.1
sympy                        1.14.0
tenacity                     9.1.2
tokenizers                   0.22.0
toml                         0.10.2
torch                        2.7.1+cu118
torchaudio                   2.7.1+cu118
torchvision                  0.22.1+cu118
tornado                      6.5.2
tqdm                         4.67.1
transformers                 4.56.0
typing_extensions            4.15.0
typing-inspection            0.4.1
tzdata                       2025.2
uritemplate                  4.2.0
urllib3                      2.5.0
watchdog                     6.0.0
wcwidth                      0.2.13
websockets                   15.0.1
wikipedia                    1.4.0